+++ lib/modules/storage/storage_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: storage
-// Updated at: 2026-05-02 16:22:35
-// Commit: feat(storage): add bi-directional event bus handler for storage
+// Updated at: 2026-05-02 17:20:52
+// Commit: fix(storage): prevent double tap duplicate action in storage
 
 class storageComponentState {
-  final String commitRef = '791';
-  final String lastMessage = 'feat(storage): add bi-directional event bus handler for storage';
-  final DateTime timestamp = DateTime.parse('2026-05-02T16:22:35+05:00');
+  final String commitRef = '794';
+  final String lastMessage = 'fix(storage): prevent double tap duplicate action in storage';
+  final DateTime timestamp = DateTime.parse('2026-05-02T17:20:52+05:00');
 }