+++ CHANGELOG.md
@@ -795,3 +795,4 @@
 - [2026-05-02 16:22] feat(storage): add bi-directional event bus handler for storage
 - [2026-05-02 16:26] perf(state): apply RepaintBoundary on animated state widgets
 - [2026-05-02 17:10] feat(ui): implement responsive ui view component
+- [2026-05-02 17:20] fix(storage): prevent double tap duplicate action in storage